I have added a Save function.  The instructor simply used a "dump" command.  It may be helpful to follow where this dump comes from, and how it operates.
For any real I/O, there is need for structure to make such operations determinate.

The instructor also recommended establishing the 5 steps in the Button action closure.  The button object initializer has two closures.  The second closure give the ability to define a label.   This label is a container for other UI elements.  The notion of block local structure ensures that such elements stay within the container.

Lastly, we add a text element in the vertical stack to supply a count of the notes in the notes array.

All of this is documented in Algorithms' subsection of Get Familiar with Apple Watch Input Interface.  The thing missing is the approach to making external devices provide their keyboard to the Watch.  This is a little more valuable for appliances like the Apple TV, but I can see some value for the watch.   I see more value in terms of the watch of shared apps  where they can operate independent and yet work together when proximity provides an adequate channel.