Here we have documentation on how the instructor thinks we should have a module method do the job of providing updates between two properties.   I miss the listeners that were part of Cocoa and Java WebObjects.   The concept of something listening for when a get of set is call is powerful feature.    We have it, sort of in the Slider shown also in SettingsView.  However, this is not the same.  The slider is part of the view.   The model is not the view, and this is where listening has to occur.   That listening in the model can advise the controller to take action.  But, that does not seem to be the pattern with SwiftUI.    Maybe, the interaction with Objective-C might make this different.
Also, we see the inclusion of an "AppStorage" declaration.  This is a descriptor to tell the compiler that the User Defaults are being accessed.    This occurs in Content View and SettingsView.   I don't have a problem with the use of User Defaults.  This has been a part of macOS programming since it inherited its features from NeXT.    Just be careful.  This is a costly global.

The documentation describes this update function and how we use User Defaults and internal values to play this kind of listen and change user defaults.  If this is a User Defaults, what happened to the User Default listeners?

Well, now is time to commit this to my GitHub.   It is a lesson, not a self derived product.  But, lessons are good things, even as refreshers of what has changed.  Such things remind us that our previous knowledge can be broken by new states of affair.